Fixed bug where init_election set acceptVal
Previously, when a node started an election, it would set the acceptVal to an empty list. The problem with this is that if this node becomes part of the quorum for another node's prepare phase, that other node will piggyback off of this acceptVal. The blockchain will eventually have an empty list as a block.